+++ api/src/services/verification-service.ts
@@ -4,7 +4,7 @@ import { SsiService } from './ssi-service';
 import { UserService } from './user-service';
 import * as KeyCollectionDb from '../database/key-collection';
 import * as VerifiableCredentialsDb from '../database/verifiable-credentials';
-import * as IdentityDocsDb from '../database/identity-docs';
+import * as IdentityDocsDb from '../database/identity-keys';
 import * as TrustedRootsDb from '../database/trusted-roots';
 import { JsonldGenerator } from '../utils/jsonld';
 import { Subject } from '../models/types/verification';
